refactor: apis -> interfaces로 폴더 이름 변경
